Delete function
delete function get the parameters root node and the isbn no(k). if the
isbn no of the root matches with the k it assign left child node to the
lt and right child to the r variables.
if lt and the rt variables are null, return null.
if lt is null lt is assign to the p ,else rt is null rt is assign to the
p and return .if none of them are not null rt value is assigned p2 and
p.