Search function
create a boolean variable found & by default it is set to false. create
a while loop to traverse the tree to search the value in the tree.. here
conditions to be checking are whether root is not null and search value
is not false. create a variable  rval to get the value of the root.